Add dialog boxes for Serene's response
